requirements.txt
amqp==1.4.9
anyjson==0.3.3
appdirs==1.4.3
billiard==3.3.0.23
celery==3.1.26.post2
dj-database-url==0.4.2
Django==1.11.7
django-appconf==1.0.2
django-autocomplete-light==3.2.10
django-bootstrap3==8.2.1
django-bootstrap4==0.0.4
django-celery==3.2.2
django-chartit==0.2.9
django-crispy-forms==1.7.0
django-discover-runner==1.0
django-easy-pdf==0.1.1
django-el-pagination==3.1.0
django-endless-pagination==2.0
django-imagekit==4.0.2
django-js-reverse==0.7.3
django-role-permissions==2.1.0
django-selectable==1.0.0
django-simple-history==1.9.0
djangorestframework==3.7.3
gitdb==0.6.4
gitdb2==2.0.3
GitPython==2.1.7
gunicorn==19.7.1
html5lib==1.0b8
kombu==3.0.37
MySQL-python==1.2.5
mysqlclient==1.3.13
olefile==0.44
packaging==16.8
pilkit==2.0
Pillow==4.3.0
pisa==3.0.33
pyparsing==2.2.0
PyPDF2==1.26.0
python-dateutil==2.6.1
pytz==2017.3
reportlab==3.2.0
requests==2.9.1
six==1.10.0
smmap==0.9.0
smmap2==2.0.3
South==0.7.3
Unidecode==0.4.21
vine==1.1.4
webencodings==0.5.1
xhtml2pdf==0.0.6
xlrd==1.1.0
xlwt==1.3.0
xmltodict==0.11.0
